# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chat Atendimento API",
    "version": "1.0.0",
    "description": "\n        API REST para sistema de chat de atendimento em tempo real.\n        \n        ## Funcionalidades\n        - Autenticação JWT\n        - Chat em tempo real com WebSocket\n        - Upload de arquivos (AWS S3)\n        - Gerenciamento de atendentes\n        - Fila de atendimento\n        - Dashboard administrativo\n        - Histórico e análises\n        \n        ## Ambientes\n        - **Desenvolvimento**: http://localhost:5000\n        - **Staging**: http://52.90.17.204:5000\n        - **Produção**: https://api.brsi.com.br\n        \n        ## Autenticação\n        Todas as rotas (exceto login e register) requerem autenticação via JWT.\n        Envie o token no header: Authorization: Bearer {token}\n      ",
    "contact": {
      "name": "BR Sistemas",
      "email": "suporte@brsi.com.br",
      "url": "https://brsi.com.br"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://brsi.com.br/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Servidor de Desenvolvimento (Local)"
    },
    {
      "url": "http://52.90.17.204:5000/api",
      "description": "Servidor de Staging (Testes)"
    },
    {
      "url": "https://api.brsi.com.br/api",
      "description": "Servidor de Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Digite o token JWT obtido no login"
      }
    },
    "schemas": {
...
      "Chat": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID único da conversa",
            "example": "60d0fe4f5311236168a109cb"
          },
          "subject": {
            "type": "string",
            "description": "Assunto da conversa",
            "example": "Dúvida sobre produto"
          },
          "customer": {
            "type": "object",
            "description": "Dados do cliente",
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string",
                "example": "Maria Santos"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "maria@example.com"
              }
            }
          },
          "agent": {
            "type": "string",
            "description": "ID do agente atribuído",
            "example": "60d0fe4f5311236168a109ca"
          },
          "status": {
            "type": "string",
            "enum": [
              "waiting",
              "active",
              "closed"
            ],
            "description": "Status atual da conversa",
            "example": "active"
          },
          "department": {
            "type": "string",
            "description": "Departamento responsável",
            "example": "vendas"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "description": "Avaliação do atendimento",
            "example": 5
          },
          "ratingComment": {
            "type": "string",
            "description": "Comentário da avaliação",
            "example": "Atendimento excelente!"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de encerramento"
          },
          "lastActivity": {
            "type": "string",
            "format": "date-time",
            "description": "Última atividade na conversa"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID único da mensagem"
          },
          "conversationId": {
            "type": "string",
            "description": "ID da conversa"
          },
          "sender": {
            "type": "string",
            "description": "ID do remetente"
          },
          "senderType": {
            "type": "string",
            "enum": [
              "customer",
              "agent",
              "system"
            ],
            "description": "Tipo do remetente"
          },
          "content": {
            "type": "string",
            "description": "Conteúdo da mensagem",
            "example": "Olá, como posso ajudar?"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "file",
              "audio"
            ],
            "description": "Tipo de mensagem",
            "example": "text"
          },
          "file": {
            "type": "object",
            "description": "Arquivo anexado (se houver)",
            "properties": {
              "url": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "size": {
                "type": "number"
              },
              "mimetype": {
                "type": "string"
              }
            }
          },
          "readBy": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "readAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "description": "Usuários que leram a mensagem"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de envio"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/chat/conversations": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Criar nova conversa",
        "description": "Cria uma nova conversa de chat. Clientes podem criar conversas que entram na fila de atendimento.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subject": {
                    "type": "string",
                    "example": "Dúvida sobre produto"
                  },
                  "initialMessage": {
                    "type": "string",
                    "example": "Olá, gostaria de saber mais sobre o produto X"
                  },
                  "department": {
                    "type": "string",
                    "example": "vendas"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conversa criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      },
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Listar conversas",
        "description": "Retorna lista de conversas do usuário. Clientes veem apenas suas próprias conversas, agentes veem conversas atribuídas ou na fila.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "waiting",
                "active",
                "closed"
              ]
            },
            "description": "Filtrar por status da conversa"
          },
          {
            "in": "query",
            "name": "assignedAgentId",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por ID do agente atribuído"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Número da página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Itens por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de conversas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Chat"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/api/chat/conversations/{id}/messages": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Buscar mensagens de uma conversa",
        "description": "Retorna todas as mensagens de uma conversa específica",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da conversa"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Número da página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Mensagens por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de mensagens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID inválido"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Conversa não encontrada"
          }
        }
      }
    },
    "/api/chat/conversations/{id}/assign": {
      "patch": {
        "tags": [
          "Chat"
        ],
        "summary": "Atribuir conversa a um agente",
        "description": "Atribui uma conversa a um agente específico (Apenas administradores)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da conversa"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "agentId"
                ],
                "properties": {
                  "agentId": {
                    "type": "string",
                    "example": "60d0fe4f5311236168a109ca"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conversa atribuída com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "403": {
            "description": "Sem permissão - Apenas administradores"
          },
          "404": {
            "description": "Conversa ou agente não encontrado"
          }
        }
      }
    },
    "/api/chat/conversations/{id}/accept": {
      "patch": {
        "tags": [
          "Chat"
        ],
        "summary": "Aceitar conversa",
        "description": "Agente aceita uma conversa da fila de atendimento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da conversa"
          }
        ],
        "responses": {
          "200": {
            "description": "Conversa aceita com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "400": {
            "description": "Erro ao aceitar conversa"
          },
          "403": {
            "description": "Sem permissão - Apenas agentes e administradores"
          },
          "404": {
            "description": "Conversa não encontrada"
          }
        }
      }
    },
    "/api/chat/conversations/{id}/close": {
      "patch": {
        "tags": [
          "Chat"
        ],
        "summary": "Fechar conversa",
        "description": "Encerra uma conversa de atendimento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da conversa"
          }
        ],
        "responses": {
          "200": {
            "description": "Conversa fechada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "400": {
            "description": "Conversa já está encerrada"
          },
          "404": {
            "description": "Conversa não encontrada"
          }
        }
      }
    },
    "/api/chat/conversations/{id}/rate": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Avaliar conversa",
        "description": "Cliente avalia o atendimento recebido",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da conversa"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "rating"
                ],
                "properties": {
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 5
                  },
                  "comment": {
                    "type": "string",
                    "example": "Atendimento excelente!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avaliação registrada com sucesso"
          },
          "403": {
            "description": "Apenas o cliente da conversa pode avaliar"
          },
          "404": {
            "description": "Conversa não encontrada"
          }
        }
      }
    },
...
  }
}